+++ service/src/main/java/uk/gov/hmcts/reform/calendar/client/BankHolidaysApi.java
@@ -3,9 +3,9 @@ package uk.gov.hmcts.reform.calendar.client;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import uk.gov.hmcts.reform.calendar.model.BankHolidays;
-import uk.gov.hmcts.reform.fpl.config.FeignConfiguration;
+import uk.gov.hmcts.reform.fpl.config.feign.FeignClientConfiguration;
 
-@FeignClient(name = "bank-holidays-api", url = "${bankHolidays.api.url}", configuration = FeignConfiguration.class)
+@FeignClient(name = "bank-holidays-api", url = "${bankHolidays.api.url}", configuration = FeignClientConfiguration.class)
 public interface BankHolidaysApi {
 
     @GetMapping(value = "/bank-holidays.json")
