+++ service/src/main/java/uk/gov/hmcts/reform/fnp/client/PaymentApi.java
@@ -5,12 +5,12 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import uk.gov.hmcts.reform.fnp.model.payment.CreditAccountPaymentRequest;
-import uk.gov.hmcts.reform.fpl.config.FeignConfiguration;
+import uk.gov.hmcts.reform.fpl.config.feign.FeignClientConfiguration;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi.SERVICE_AUTHORIZATION;
 
-@FeignClient(name = "payment-api", url = "${payment.api.url}", configuration = FeignConfiguration.class)
+@FeignClient(name = "payment-api", url = "${payment.api.url}", configuration = FeignClientConfiguration.class)
 public interface PaymentApi {
     @PostMapping("/credit-account-payments")
     CreditAccountPaymentRequest createCreditAccountPayment(
