+++ service/src/main/java/uk/gov/hmcts/reform/fpl/config/feign/GlobalFeignConfiguration.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class FeignConfiguration {
+public class GlobalFeignConfiguration {
 
     @Bean
     public Retryer retryer() {
