+++ service/src/main/java/uk/gov/hmcts/reform/rd/client/OrganisationApi.java
@@ -4,7 +4,7 @@ import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
-import uk.gov.hmcts.reform.fpl.config.FeignConfiguration;
+import uk.gov.hmcts.reform.fpl.config.feign.FeignClientConfiguration;
 import uk.gov.hmcts.reform.rd.model.Organisation;
 import uk.gov.hmcts.reform.rd.model.OrganisationUser;
 import uk.gov.hmcts.reform.rd.model.OrganisationUsers;
@@ -13,7 +13,7 @@ import uk.gov.hmcts.reform.rd.model.Status;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi.SERVICE_AUTHORIZATION;
 
-@FeignClient(name = "rd-professional-api", url = "${rd_professional.api.url}", configuration = FeignConfiguration.class)
+@FeignClient(name = "rd-professional-api", url = "${rd_professional.api.url}", configuration = FeignClientConfiguration.class)
 public interface OrganisationApi {
     @GetMapping("/refdata/external/v1/organisations/users")
     OrganisationUsers findUsersInOrganisation(
